Move the touchmove handler to the body; prevent unwanted double-slap tap
Putting touchmove on the body means that a touch can begin outside of the cat and move onto the cat to perform a slap.
Before, you had to start on the cat. This also prevents a touch that starts on the cat from causing a move-based slap
until it moves off and back onto the cat. This fixes most of the double-slappage.