Refuse to slap the cat until a click or tap is used to slap
My browser refuses to play sound until a click/tap/keyboard/etc event occurs. To avoid silent slaps, movement
based slaps are ignored until a tap or click is used to cause a slap. This also fixes a weird bug where
the first tap on the cat on a phone registers as a mousemove event, causing a doubleslap.

The web is great.